requirements.txt
Django==5.2.6
Pillow==10.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.32.3
django-allauth==65.11.2
PyJWT==2.10.1
cryptography==45.0.7
python-decouple==3.8
